# docs/api.yaml
openapi: 3.1.0
...
info:
  title: MBS Backend
  x-logo:
    url: "https://github.com/jaxcksn/jaxcksn/blob/main/files/ttu_cs_light.png?raw=true"
    backgroundColor: "#FFFFFF"
    altText: Texas Tech Department of Computer Science
  version: 0.0.1
  summary: MBS API Contract
  description: This defines all the endpoints and routes that should exist in the API. This is meant to be a rough guide and may change as needed.
servers:
  - url: "http://localhost:5050"
    description: Local
  - url: "https://api.jaxcksn.dev"
    description: Staging
paths:
...
  /admin/showing:
    post:
      summary: Create A Showing (Admin)
      tags: []
      responses:
        "201":
          description: Created
        "403":
          description: Forbidden
      operationId: create-showing
      x-stoplight:
        id: st74wtozns73t
      description: Creates a new movie and showing.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - showing_start
                - showing_end
                - price
                - poster_url
                - times
              properties:
                title:
                  type: string
                  x-stoplight:
                    id: 7jt8lxpq0j1oe
                description:
                  type: string
                  x-stoplight:
                    id: ax5tbs0cz1ddr
                showing_start:
                  type: string
                  x-stoplight:
                    id: y7z8woqjcfxqp
                  format: date
                showing_end:
                  type: string
                  x-stoplight:
                    id: fzepav6x9iu7w
                  format: date
                price:
                  type: number
                  x-stoplight:
                    id: 71aemdzvwzken
                poster_url:
                  type: string
                  x-stoplight:
                    id: izi8zajqsbnkx
                times:
                  type: array
                  x-stoplight:
                    id: hml6xu865mnxo
                  items:
                    x-stoplight:
                      id: s2pqn8jwmvo6q
                    type: string
      security:
        - BEARER: []
    parameters: []
    patch:
      summary: Modify a Showing (Admin)
      tags: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
      operationId: patch-showing
      x-stoplight:
        id: hyjket3opapkw
      description: Modifies a showing/movie.
      parameters:
        - schema:
            type: string
          in: query
          name: id
          description: The ID of the showing to modify
      security:
        - BEARER: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  x-stoplight:
                    id: szka1vq28uco8
                description:
                  type: string
                  x-stoplight:
                    id: 7p9e3chhadzhw
                release_date:
                  type: string
                runtime:
                  type: string
                cast:
                  type: string
                showing_start:
                  type: string
                  x-stoplight:
                    id: 5rmmhnqkt76fs
                showing_end:
                  type: string
                  x-stoplight:
                    id: ce5d4shrqzxxx
                price:
                  type: number
                  x-stoplight:
                    id: xr1rai0xw2grh
                poster_url:
                  type: string
                  x-stoplight:
                    id: 6ynqa7dak1v0w
                times:
                  type: array
                  x-stoplight:
                    id: 2y2b7yioyyydi
                  items:
                    x-stoplight:
                      id: 9x06mwnxfujtg
                    type: string
        description: "If a property in the body is left blank, the API will not modify it."
    delete:
      summary: Delete a Showing (Admin)
      tags: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
      operationId: delete-showing
      x-stoplight:
        id: dt0sot93eu44g
      description: Removes a showing/movie
      security:
        - BEARER: []
      parameters:
        - schema:
            type: string
          in: query
          name: id
          description: The ID of the movie to delete
  /admin/showings:
    get:
      summary: Fetch All Showings (Admin)
      tags: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  x-stoplight:
                    id: lt16i55a0mqqf
                  type: object
                  required:
                    - movie
                    - id
                    - release_date
                    - price
                  properties:
                    movie:
                      type: object
                      x-stoplight:
                        id: noxv1um1868sl
                      required:
                        - title
                        - poster_url
                      properties:
                        title:
                          type: string
                          x-stoplight:
                            id: ndnyq27b7vj8d
                        poster_url:
                          type: string
                          x-stoplight:
                            id: 4ksjew59fuvx2
                    id:
                      type: string
                      x-stoplight:
                        id: toa7w6zh60kdv
                    release_date:
                      type: string
                      x-stoplight:
                        id: zlxdlggosnkxp
                    price:
                      type: number
                      x-stoplight:
                        id: uruogqaedo5fp
        "403":
          description: Forbidden
      operationId: get-admin-showings
      x-stoplight:
        id: hre2b8n9e1ioi
      security:
        - BEARER: []
      description: Gets all the showings for an admin
...
components:
...
  securitySchemes:
    BEARER:
      type: http
      scheme: bearer